+++ R/tw_get_property_with_details.R
@@ -10,10 +10,18 @@
 #' @examples
 #' # Get "female form of label", including language
 #' tidywikidatar:::tw_get_property_with_details_single(id = "Q64733534", p = "P2521")
-tw_get_property_with_details_single <- function(id, p) {
-  item <- tryCatch(tw_get_item(id = id), error = function(e) {
-    as.character(e[[1]])
-  })
+tw_get_property_with_details_single <- function(
+  id,
+  p,
+  retry = 10,
+  user_agent = tidywikidatar::tw_get_user_agent()
+) {
+  item <- tryCatch(
+    tw_get_item(id = id, retry = retry, user_agent = user_agent),
+    error = function(e) {
+      as.character(e[[1]])
+    }
+  )
 
   if (is.character(item)) {
     cli::cli_alert_danger(item)
@@ -30,9 +38,12 @@ tw_get_property_with_details_single <- function(id, p) {
   ) {
     id <- item %>%
       purrr::pluck(1, "redirect")
-    item <- tryCatch(tw_get_item(id = id), error = function(e) {
-      as.character(e[[1]])
-    })
+    item <- tryCatch(
+      tw_get_item(id = id, retry = retry, user_agent = user_agent),
+      error = function(e) {
+        as.character(e[[1]])
+      }
+    )
   }
 
   claims <- item %>% purrr::pluck(1, "claims")
@@ -61,7 +72,13 @@ tw_get_property_with_details_single <- function(id, p) {
 #' @examples
 #' # Get "female form of label", including language
 #' tw_get_property_with_details(id = "Q64733534", p = "P2521")
-tw_get_property_with_details <- function(id, p, wait = 0) {
+tw_get_property_with_details <- function(
+  id,
+  p,
+  wait = 0,
+  retry = 10,
+  user_agent = tidywikidatar::tw_get_user_agent()
+) {
   pb <- progress::progress_bar$new(total = length(unique(id)))
   purrr::map2_dfr(
     .x = unique(id),
@@ -71,7 +88,9 @@ tw_get_property_with_details <- function(id, p, wait = 0) {
       Sys.sleep(wait)
       tw_get_property_with_details_single(
         id = id,
-        p = p
+        p = p,
+        retry = retry,
+        user_agent = user_agent
       )
     }
   )